debt - move minimalTranslations to where it is used
