+++ src/tsconfig.strictNullChecks.json
@@ -317,7 +317,6 @@
 		"./vs/platform/menubar/electron-main/menubar.ts",
 		"./vs/platform/menubar/electron-main/menubarService.ts",
 		"./vs/platform/menubar/node/menubarIpc.ts",
-		"./vs/platform/node/minimalTranslations.ts",
 		"./vs/platform/node/package.ts",
 		"./vs/platform/node/product.ts",
 		"./vs/platform/node/test/zip.test.ts",
@@ -559,6 +558,7 @@
 		"./vs/workbench/contrib/files/electron-browser/views/explorerDecorationsProvider.ts",
 		"./vs/workbench/contrib/localizations/electron-browser/localizations.contribution.ts",
 		"./vs/workbench/contrib/localizations/electron-browser/localizationsActions.ts",
+		"./vs/workbench/contrib/localizations/electron-browser/minimalTranslations.ts",
 		"./vs/workbench/contrib/logs/common/logConstants.ts",
 		"./vs/workbench/contrib/logs/electron-browser/logs.contribution.ts",
 		"./vs/workbench/contrib/logs/electron-browser/logsActions.ts",
