+++ src/vs/workbench/contrib/localizations/electron-browser/minimalTranslations.ts
